requirements.txt
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/pytest-sina-framework.git@master#egg=pytest-sina-framework
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/pytest-rst-report.git@master#egg=pytest_rst_report
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/restlib.git@master#egg=restlib
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/clilib.git@master#egg=clilib
git+ssh://git@gitlab.sinacomsys.local/quality-assurance/test-infrastructure/sshlib.git@master#egg=sshlib
pytest_check==2.1.4
colorama

